messy version, but should work better on mobile
Lots of features scrapped, most importantly, prevent closing of popups by clicking on the map. This should make the survey much more usable on mobile phones because popup dialog doesn't dissappear randomly and the survey registers users' taps much better. Some problems persist, such as save changes and delete these records sometimes losing their "focus" and becoming unclickable. Though, x button still works. This is very promising and this has enabled me to completely remove several features out of the survey, such as the ugly auxiliary variable apu is no longer needes and I have no use for the bottom Polygon which used to catch user's clicks in the sea in the research area. All of them are commented in this version, but will be deleted in the next.